Windows build fixes.
svn path=/trunk/libavg/; revision=4163

+++ src/base/Point.cpp
@@ -254,7 +254,7 @@ bool almostEqual(const DPoint& pt1, const DPoint& pt2)
 template<class NUM>
 double calcDist(const Point<NUM>& pt1, const Point<NUM>& pt2)
 {
-    return sqrt(sqr(pt1.x-pt2.x)+sqr(pt1.y-pt2.y));
+    return sqrt(double(sqr(pt1.x-pt2.x)+sqr(pt1.y-pt2.y)));
 }
 
 template<class NUM>
